+++ packages/get_hooked/lib/widgets.dart
@@ -2,3 +2,4 @@ export 'widgets/clipped_decorated_box.dart';
 export 'widgets/hook_widgets.dart';
 export 'widgets/immutable_subtree.dart';
 export 'widgets/ref_paint.dart';
+export 'widgets/render_get.dart';