+++ packages/get_hooked/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.3.0
 - Decided that `RenderHookElement` is too convoluted!
   - API has been reworked into `RefPaint`
+- A bunch of benchmarking!
+- `RenderGet`, a super-performant render object widget
 
 
 ## 0.2.2